+++ app/scripts/lib/extension-instance.js
@@ -42,6 +42,12 @@ function Extension () {
     } catch (e) {}
 
     try {
+      if (browser[api]) {
+        _this[api] = browser[api]
+      }
+    } catch (e) {}
+
+    try {
       _this.api = browser.extension[api]
     } catch (e) {}