+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Current Master
 
+- Changed transaction approval from notifications system to popup system.
 - Forms now retain their values even when closing the popup and reopening it.
 
 ## 2.9.2 2016-08-24